frontend/pyproject.toml
[tool.poetry]
name = "front"
version = "0.1.0"
description = ""
authors = ["SNMHZ <snmhz325@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
streamlit = "^1.9.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
